float

+++ calc/Main.java
@@ -4,19 +4,19 @@ public class Main {
         Scanner in = new Scanner(System.in);
 
         System.out.println("Please input first number:");
-        int num1 = in.nextInt();
-        System.out.printf("Your input %d as a first number. \n", num1);
+        float num1 = in.nextFloat();
+        System.out.printf("Your input %f as a first number. \n", num1);
         System.out.println("Please input second number:");
-        int num2 = in.nextInt();
-        System.out.printf("Your input %d as a second number. \n", num2);
-        long sum = num1+num2;
-        System.out.printf("Your \nSum is %d \n", sum);
-        long dif = num1-num2;
-        System.out.printf("Diference is %d \n", dif);
-        long mul = num1*num2;
-        System.out.printf("Multiplication is %d \n", mul);
+        float num2 = in.nextFloat();
+        System.out.printf("Your input %f as a second number. \n", num2);
+        double sum = num1+num2;
+        System.out.printf("Your \nSum is %f \n", sum);
+        double dif = num1-num2;
+        System.out.printf("Diference is %f \n", dif);
+        double mul = num1*num2;
+        System.out.printf("Multiplication is %f \n", mul);
         if (num2 != 0) {
-            float div = (float)num1 / num2;
+            float div = num1 / num2;
             System.out.printf("Division is %f \n", div);
         } else {
             System.out.print("Division impossible");
